Labs/CodeQL-Demo-App/requirements.txt
flask==1.0         # CVE-2018-1000656
pyyaml==4.2b1      # CVE-2017-18342, CVE-2020-1747
requests==2.19.1   # CVE-2018-18074